Move BorrowedValue to value module
