+++ src/lib.rs
@@ -16,6 +16,6 @@ pub mod ast;
 pub mod lexer;
 lalrpop_util::lalrpop_mod!(pub parser, "/corn.rs");
 
-pub use de::{from_str, parse, BorrowedValue, Deserializer};
+pub use de::{from_str, parse, Deserializer};
 pub use error::{Error, Result};
-pub use value::{Integer, Object, Value};
+pub use value::{BorrowedValue, Integer, Object, Value};
